refactor(VBanner): move children back to main component
after process review, the abstraction was too much for
this component, but will apply in the future.
