--- packages/vuetify/src/components/VBanner/VBannerContent.scss
@@ -1,21 +1,0 @@
-// Imports
-@import './index';
-
-.v-banner-content {
-  align-items: center;
-  display: flex;
-  flex: 1 1 auto;
-  line-height: $banner-content-line-height;
-  padding-bottom: $banner-content-padding-bottom;
-
-  &--is-mobile {
-    padding-bottom: $banner-mobile-content-padding-bottom;
-  }
-
-  &--single-line {
-    padding-bottom: $banner-single-line-content-padding-bottom;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-}
