--- packages/vuetify/src/components/VBanner/VBannerThumbnail.scss
@@ -1,48 +1,0 @@
-// Imports
-@import './index';
-
-.v-banner-thumbnail {
-  align-self: flex-start;
-  border-radius: $banner-thumbnail-border-radius;
-  flex: 0 0 auto;
-  height: $banner-thumbnail-height;
-  width: $banner-thumbnail-width;
-
-  @include ltr () {
-    margin-left: $banner-thumbnail-margin-start;
-    margin-right: $banner-thumbnail-margin-end;
-  }
-
-  @include rtl () {
-    margin-right: $banner-thumbnail-margin-start;
-    margin-left: $banner-thumbnail-margin-end;
-  }
-
-  &--is-mobile {
-    @include ltr() {
-      margin-right: $banner-mobile-thumbnail-padding-start;
-    }
-
-    @include rtl() {
-      margin-left: $banner-mobile-thumbnail-padding-start;
-    }
-  }
-}
-
-.v-banner-thumbnail__avatar {
-  border-radius: inherit;
-  height: inherit;
-  width: inherit;
-}
-
-.v-banner-thumbnail__icon {
-  align-items: center;
-  background-color: rgb(var(--v-theme-primary));
-  border-radius: inherit;
-  color: $banner-thumbnail-icon-color;
-  display: inline-flex;
-  font-size: $banner-thumbnail-icon-font-size;
-  height: inherit;
-  width: inherit;
-  justify-content: center;
-}
